% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{warwick}
\alias{warwick}
\title{Warwick Mountain Lithgeochemistry}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1863 rows and 44 columns.}
\source{
DP ME 505 , Version 1, 2018. Lithogeochemical Data from the Warwick Mountain
Area, Eastern Cobequid Highlands, Nova Scotia, by T.G. MacHattie and C.C.
MacMullen. Nova Scotia Department of Energy and Mines.
\url{https://novascotia.ca/natr/meb/download/dp505.asp}
}
\usage{
warwick
}
\description{
This dataset is a data frame of rock samples collected in multiple
field seasons by the Nova Scotia Department of Energy and Mines in the
Warwick Mountain area of Nova Scotia, Canada. It contains qualitative
information about rock types, as well as pXRF-measured geochemistry.
}
\keyword{datasets}
